updating with validations, testing in postman, graphql, mongoDB atlas and finalizing submission

+++ COMP3133_100627702_Assignment1/schema/graphqlSchema.js
@@ -4,7 +4,7 @@ const jwt = require('jsonwebtoken')
 const User = require('../models/User')
 const Employee = require('../models/Employee')
 
-module.exports = buildSchema(`
+const graphqlSchema = buildSchema(`
 
     type User {
 
@@ -50,9 +50,9 @@ module.exports = buildSchema(`
             last_name: String!,
             email: String,
             gender: String,
-            destignation: String!,
+            designation: String!,
             salary: Float!,
-            date_of_joinig: String!,
+            date_of_joining: String!,
             department: String!,
             employee_photo: String
         ): Employee
@@ -62,9 +62,9 @@ module.exports = buildSchema(`
             last_name: String!,
             email: String,
             gender: String,
-            destignation: String!,
+            designation: String!,
             salary: Float!,
-            date_of_joinig: String!,
+            date_of_joining: String!,
             department: String!,
             employee_photo: String       
         ): Employee
@@ -74,7 +74,17 @@ module.exports = buildSchema(`
 
 const queryHandlers = {
 
+
+// -- Resolvers and Validations --
     login: async ({ username, password }) => {
+
+        if (!username || username.trim().length === 0){
+            throw new Error("USERNAME IS REQUIRED.")
+        }
+
+        if (!password || password.trim().length === 0){
+            throw new Error("PASSWORD IS REQUIRED.")
+        }
 
         const user = await User.findOne({ username })
         
@@ -89,6 +99,23 @@ const queryHandlers = {
 
     signup: async ({ username, email, password }) => {
 
+        if (!username || username.trim().length < 4){
+            throw new Error("USERNAME MUST BE 4 CHARACTERS.")
+        }
+
+        if (!email || !email.includes("@") || !email.includes(".")) {
+            throw new Error("INVALID. MUST HAVE @")
+        }
+
+        if (!password || password.length < 6){
+            throw new Error("PASSWORD MUST BE > 6 CHARS.")
+        }
+
+        const userExists = await User.findOne({ Sor: [{ username }, { email }] })
+        if(userExists) {
+            throw new Error("USER ALREADY EXISTS.")
+        }
+
         const hashedPass = await bcrypt.hash(password, 12)
 
         const user = new User({ username, email, password: hashedPass })
@@ -99,8 +126,8 @@ const queryHandlers = {
 
     getAllEmployees: async () => await Employee.find(),
 
-    searchAllEmployeesByEid: async () => await Employee.findOne({ eid }),
-    
+    searchEmployeesByEid: async ({ eid }) => await Employee.findById(eid ),
+
     searchEmployees: async ({ designation, department }) => {
 
         const query = {}
@@ -115,6 +142,41 @@ const queryHandlers = {
 
     addEmployee: async (args) => {
 
+        if(!first_name || first_name.trim().length < 1){
+            throw new Error("FIRST NAME MUST BE LONGER THAN 1 CHARS.")
+        }
+
+        if(!last_name || last_name.trim().length < 1){
+            throw new Error("FIRST NAME MUST BE LONGER THAN 1 CHARS.")
+        }
+
+        if(!last_name || last_name.trim().length < 1){
+            throw new Error("LAST NAME MUST BE LONGER THAN 1 CHARS.")
+        }
+
+        if(!designation || designation.trim().length < 1){
+            throw new Error("DESIGNATION MUST BE LONGER THAN 1 CHARS.")
+        }
+
+        if(!department || department.trim().length < 1){
+            throw new Error("DEPARTMENT MUST BE LONGER THAN 1 CHARS.")
+        }
+
+        if(!salary || salary < 1000){
+            throw new Error("SALARY MUST BE > 1000.")
+        }
+
+        if (email){
+            if (!email.includes("@") || !email.includes(".")){
+                throw new Error("INVALID. MUST HAVE @")
+            }
+
+            const employeeExists = await Employee.findOne({ email });
+            if (employeeExists){
+                throw new Error("EMPLOYEE EXISTS.")
+            }
+        }
+
         const employee = new Employee(args)
         
         return await employee.save()
@@ -123,11 +185,41 @@ const queryHandlers = {
 
     updateEmployeeByEid: async ({ eid, ...args }) => {
 
+        const employee = await Employee.findById(eid)
+        if (!employee) throw new Error("EMPLOYEE CAN NOT BE FOUND.")
+
+        if (args.first_name && args.first_name.trim().length < 1){
+            throw new Error("FIRST NAME MUST BE LONGER THAN 2 CHARS.")
+        }
+
+        if (args.last_name && args.last_name.trim().length < 1){
+            throw new Error("LAST NAME MUST BE LONGER THAN 2 CHARS.")
+        }
+
+        if (args.email){
+            if (!args.email.includes("@") || !args.email.includes(".")){
+                throw new Error("INVALID. MUST HAVE @")
+            }
+        }
+
+        if(args.salary && args.salary < 1000){
+            throw new Error("SALARY MUST BE > 1000.")
+        }
+
+        if(args.date_of_joining && isNaN(Date.parse(args.date_of_joining))){
+            throw new Error("INVALIDE DATE.")
+        }
+
         return await Employee.findByIdAndUpdate(eid, args, { new: true });
 
     },
 
     deleteEmployeeByEid: async ({ eid }) => {
+
+        const employee = await Employee.findByid(eid)
+        if(!employee){
+            throw new Error("EMPLOYEE NOT CAN NOT BE FOUND.")
+        }
 
         await Employee.findByIdAndDelete(eid);
         
@@ -137,4 +229,7 @@ const queryHandlers = {
 
 };
 
-module.exports.guerylHandlers = queryHandlers;
+module.exports = {
+    graphqlSchema,
+    queryHandlers,
+};
